refactors num likes and date formatting to src/utility as modules
