+++ frontend/src/components/post/Post.js
@@ -1,6 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import './Post.css';
 import getSessionUserID from '../utility/getSessionUserID';
+import formatNumLikes from '../utility/getNumLikes';
+import formatLikesUsersPreview from '../utility/getNumLikesUserPreview';
+import formatFullDateString from '../utility/getFullDateString';
+import convertRelativeDateTimeString from '../utility/getRelativeTimestamp';
+
 
 const Post = ({post}) => {
   
@@ -58,86 +63,19 @@ const Post = ({post}) => {
 // ============= DISPLAYING LIKES ==================
 
   // --------- "X like(s)" ---------------
-  const formatNumLikes = (arr) => {
-    const numLikes = arr.length;
-
-    if (numLikes === 1){
-      return "1 like"
-    } else {
-      return `${numLikes} likes`
-    }}
-
   const likes_formatted = formatNumLikes(post.likes)
 
   // --------- "You/User and X others liked this" ---------------
-  const formatLikesUsersPreview = (arr) => {
-    const numLikes = arr.length;
-    // 0 likes
-    if (numLikes === 0){
-      return "Noone liked this"
-    // 1 like
-    } else if (numLikes === 1){ //Check if sessionUserID liked this
-      return (arr.some(user => user._id === sessionUserID) ? 'You liked this' : `${arr[0].email} liked this`);
-    // 2 likes
-    } else if (numLikes === 2){
-      return (arr.some(user => user._id === sessionUserID) ? 'You and 1 other liked this' : `${arr[0].email} and 1 other liked this`);
-    // 3 or more likes
-    } else {
-      return (arr.some(user => user._id === sessionUserID) ? `You and ${numLikes - 1} others liked this` : `${arr[0].email} and ${numLikes - 1} others liked this`);
-    }
-  };
-
-  const likes_formatted_with_user_preview = formatLikesUsersPreview(post.likes)
+  const likes_formatted_with_user_preview = formatLikesUsersPreview(post.likes, sessionUserID)
 
 
 // ======== FORMATTING TIME ==============
-// TODO refactor this into a module to reuse for comments/other content
-
   const postedDateTime = new Date(post.date_posted);
-  const currentDateTime = new Date();
- 
+
   // ------------ '19 Nov 2023, 5:45PM' -------------
-  const options = {
-    day: 'numeric',
-    month: 'short',
-    year: 'numeric',
-    hour: 'numeric',
-    minute: 'numeric',
-    hour12: true,
-  };
+  const fullDateTimeString = formatFullDateString(postedDateTime)
 
-  const fullDateTimeString = postedDateTime.toLocaleString('en-GB', options);
-
-  // ------------ 'X m ago / X h ago / 1 d ago / fullDateTime --------------
-
-  const convertRelativeDateTimeString = (dateObject) => {
-    // get time difference in seconds
-    const timeDifferenceInSeconds = (currentDateTime.getTime() - postedDateTime.getTime()) / 1000;
-    const seconds = timeDifferenceInSeconds; // shorter variable name for relative timestamp use below
-
-    // show the relative timestamp
-    if (seconds < 60) { 
-      // if under 1 min, show seconds ago
-      return `${Math.floor(seconds)} seconds ago`
-
-    } else if (seconds < 3600) {
-      // if under 60 min, show minutes ago
-      return ((Math.floor(seconds/60) === 1) ? '1 minute ago' : `${Math.floor(seconds / 60)} minutes ago`);
-
-    } else if (seconds < 86400) {
-      // if under 24 hours, show hours ago
-      return ((Math.floor(seconds/3600) === 1) ? '1 hour ago' : `${Math.floor(seconds / 3600)} hours ago`);
-
-    } else if (seconds < 259200) {
-      // if less than 3 days ago, show days ago
-      return ((Math.floor(seconds/86400) === 1) ? '1 day ago' : `${Math.floor(seconds / 86400)} days ago`);
-
-    } else {
-      // else show fullDateTimeString
-      return fullDateTimeString;
-    }
-  }
-
+  // ------------ 'X seconds ago / X minutes ago / X hours ago / X days ago / fullDateTime --------------
   const relativeDateTimeString = convertRelativeDateTimeString(postedDateTime);
