+++ frontend/src/components/utility/getNumLikes.js
@@ -1,0 +1,11 @@
+  // --------- "X like(s)" ---------------
+  const formatNumLikes = (arr) => {
+    const numLikes = arr.length;
+
+    if (numLikes === 1){
+      return "1 like"
+    } else {
+      return `${numLikes} likes`
+    }}
+
+export default formatNumLikes;
